refactor: [WQ-68] JWT 프로바이더 메서드 명칭 변경 및 사용처 수정
- 기존 `createAccessToken`, `createRefreshToken` 메서드명을 `createAccessTokenDeprecated`, `createRefreshTokenDeprecated`로 변경
- 관련 메서드 호출부 코드 수정
- `getRole` 메서드 명칭을 `getRoleDeprecated`로 변경
- `JwtProperties` 및 관련 클래스 경로 수정
- `opaqueId` 필드 추가 및 관련 로직 반영
